Tweaks to fix compiler errors after cleanup
